requirements.txt
beautifulsoup4==4.12.3
numpy==1.26.4
pandas==2.2.2
pyvi==0.1.1
requests==2.31.0
scikit-learn==1.4.2
scipy==1.13.0
tqdm
langchain-community==0.3.3
lancedb==0.15.0
rank_bm25==0.2.2
sentence-transformers==3.2.1